requirements.txt
sqlalchemy
python-dotenv
PyMySQL